// generated/src/app/api-models/edi-report-work.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class EdiReportWork {
  seqPrediId: number;
  seqMessageId: number;
  messageId: number;
  reportText: string;
  seqSourceId: number;
  currentValue: string;
  newValue: string;
  reportCode: string;
  subscriberId: string;
  personNumber: string;
  lastName: string;
  firstName: string;
  columnChanged: string;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
}